Refactor the Deck component logic to it's own hook
